integration tests: use dstat to monitor system
Dstat is a tool used to monitor system in real-time. It's used by
Devstack in OpenStack gate to monitor system during jobs.
This tool would be useful for Puppet OpenStack Integration jobs so we
would be able to monitor the system during Puppet & Tempest runs, and
help us to debug eventual failures.

Change-Id: I5b6d6c662f1c23478ba0e0fb4f3d861a060abce7
Depends-On: I79c5b9521265e97595d3323149ba6e7d1e8e972b