pin tempest to keep setuping heat roles
A recent commit in Tempest [1] broke backward compatibility for Heat
testing.
stack_owner_role is not managed anymore by Tempest and it makes tests
fails.

This is a workaround until it's sorted out in Tempest.
Until this time, we pin Tempest to a previous commit so we can continue
to test scenario002.

[1] https://github.com/openstack/tempest/commit/aa461b746bac5144d99454ebdfece860e3fd7173
Related-Bug: #1543785

Change-Id: Ib19c25b6e4d9af808a3abab4d4ecbded2f33c034