repos: bump RDO to a latest repo
Since RDO did not promote current-passed-ci for a while let's bump the
repo to the latest URL available, so we can have some new packages and
test latest Mitaka.

Also it will help to have acceptance tests working on Mistral, since a
bug has been fixed in packaging but not promoted.

Change-Id: I4f8b213354d18a11c0d97250f1b88c7927790c7f