Use "default" domain in auth credentials
Even if we change the default domain, the admin user and admin project
created by keystone::roles::admin belong to the domain with ID
"default". When the admin token is disabled with
keystone::disable_admin_auth_token, puppet needs to use the admin user
and password with the domain it was created in.

Eventually the keystone::roles::admin class should be adjusted to
properly re-create the admin resources in the new default domain. This
patch should help fix CI in the meantime.

Change-Id: I6eea23c588de0498b8f3c412e3fe9c4a09efd2f7