Fix CI status on both ubuntu/centos
Both changes need to happen in the same patchset.

1/ centos/workaround: make sure Keystone installed before RabbitMQ

We currently have an issue in Puppet OpenStack CI where DIB image for
centos7 has a modified version of /etc/login.defs that allows RabbitMQ
to steal UID of OpenStack packaging (Keystone most often).

While we are investigating that, this patch is a workaround to make sure
we first install Keystone and then RabbitMQ, on centos7 systems.

2/ disable telemetry tests on Ubuntu.
UCA is currently updating Ceilometer / Aodh and tests are broken now.
Let's skip them.

Change-Id: Ibfc3988a4de47c9d7d97159e7d1c0e57d64979ae