run_tests: uninstall epel-release
epel-release is installed by OpenStack Infra when building images.
The problem is RDO + EPEL does not work find, there are some
dependencies issues.

This patch makes sure EPEL is not installed when we start our CI jobs.
The next steps will be to run puppet-openstack-integration with EPEL
disabled in ceph::repo, and patch ceph::repo to use centos ceph repos [1]

[1] https://bugs.launchpad.net/puppet-ceph/+bug/1562998

This patch is part of the efforts to stop using EPEL at all.

Change-Id: I506754e2c5021ee728675a8771f2e24b8576f7d5