functions: add defaults for ZUUL_BRANCH & ZUUL_REF
When running periodic jobs, ZUUL_BRANCH & ZUUL_REF are empty because we
don't have any reference or branch set for the jobs.
To not make zuul failing when cloning Puppet modules, let's set defaults
values:

ZUUL_BRANCH default to master (we are running tests against master now)
If we need to test stable branches, we'll revisit later but it would be
a bigger change (job names, etc).

ZUUL_REF default to None.

Change-Id: I34b298230bc174a3fdd431095043543c4c223296