deploy neutron with SSL & IPv6
Starting from Newton, VMs can be deployed with a metadata server that
runs with SSL enabled.
This patch aims to activate the tests for that, so we'll have a better
coverage of Neutron in SSL & IPv6 environment.
Note: I could not enable SSl for neutron metadata agent, I found this
bug: https://bugs.launchpad.net/neutron/+bug/1514424 and I'm not sure it
actually works.

Change-Id: Ia5b19d22549fc0c891a46cf47a742b35eaec6276