Align the stars to bump CI to latest Newton
1) scenario001: disable ceilometer tests on Ubuntu
Because we need this backport in UCA:
https://review.openstack.org/#/c/318503/
And we won't have it before... who knows?
Let's disable the tests for Ceilometer on Ubuntu until UCA provides a
packaging update.
Let's also disable Telemetry at all on Ubuntu, since Ceilometer can't be
tested until UCA does the backport.

2) add RDO Newton dependencies repository, so we have paramiko >= 2.0
and have a more stable gate.

3) Update RDO trunk URL to have a very recent URL.

4) Update nova endpoints to match with what is required upstream (2.1)

5) Un-pin Tempest, so we keep testing master (our desire in our master
gate but we pin it in stable branches).

Change-Id: I10dc1beeca3ec47fd4c3a866aa3e1d3bc9d0aa81