Allow run_tests.sh to be rerun
This updates the return code checking to allow for rerunning of the
script when the puppet fails. Previously it would error out if puppet
returned 0 when there were no changes which is not correct. This change
updates the return code check for puppet to include 0 and 2 for the
repo configuration and initial puppet run.

Change-Id: I5252b7561ea589f3f49f7f6b43269328b326aa7e