provision: require admin user before creating nova flavor
Require to first create admin keystone before trying to create nova
flavors.
The reason of such a change is that we figured when running a single
node deployment, we need to make sure to create the flavor at the right
time in the catalog otherwise we have seen some cases where the flavor
was created before having Nova API actually working.

Change-Id: I299d4c372da702232eaa7cb34b690e372f56e701