Stop using tox
This commit abandons the tox usage in run_tests.sh script. At this
point tox is only being used to create a venv and not run any tests.
The primary advantage tox gives you is maintaining separate test
environments by handling the venv management, job commands, etc. Since
we are only using it venvs, tox is a bit heavyweight when we can just
call virtualenv and pip directly. This just make it more explict so
people can see exactly what is happening when we run tempest. It also
continues the work to decouple running the tests in CI from the tempest
repo itself.

Change-Id: I57ded07a2fd2cde4d532f65e68bacbf62547cf1f