Allow p-o-i to install itself from current source when used without Zuul
puppet-openstack-integration currently takes care of installing itself.
This is done through the Puppetfile which references the p-o-i repository.

This works with zuul-cloner because Zuul with ensure the current reference
is cloned.
However, when running outside of the gate, this means that p-o-i will
completely disregard any local code and instead install from the git source
specified in the Puppetfile.

Change-Id: I665347dbede2c53bd5537d4a29d07bfe679c50e5