Update cp line to remove redundant param
The use of -r for cp command is redundant with -a being used as
documentation on Linux has -R and -r as the same parameter. The
issue on Mac OS X is that it will error on about using -R and -r
with cp. Change removes redundant -r from cp command.

Change-Id: If7122efc8b84f8aa8a07fe4f8ee7568a048c9461