[Newton] Not all Puppet modules have a stable/newton branch
When inside the gate, the modules are really installed through
zuul-cloner which falls back to master automatically if the specified
branch does not exist.

When outside the gate, however, r10k will fail on modules if the
stable/newton branch does not exist.
For the modules that do not have a stable/newton branch, let's set
them to master.

Change-Id: I422a4c24b51a4d7abe16b97acac5c0b9936ab5ab