Split out puppet logs
To improve troubleshooting and readability, let's not have the puppet
logs print to the console.  This change updates the puppet apply to
output the logs to puppet.log which will still be captured and provided
in the CI results.  When the puppet apply fails, we will grep for
warnings and errors from the puppet log and only print those out which
should simplify troubleshooting failures.

Change-Id: I1084223f805f1960aebb9c5868fdef84ff3667e4
(cherry picked from commit e4c6f48efa90bab940d30118be25e4849c4aea5e)