Make version of tempest configurable
This patch makes tempest version configurable so we can override
it in CI for DLRN promotion jobs for stable branches.

There have been some backwards incompatible changes in tempest
and ceilometer ( https://review.openstack.org/#/c/395506/ ) not
tagged yet in stable/newton. In order to get p-o-i pass with this
commits we need to use an updated version of tempest.

Change-Id: Ifb37a5a3a809c8a94d055d9cdc2b880e1ec9a115