Align stars to fix CI
Currently we are suffering from a flakey cinder test due to an issue
with the RDO provided PyMySQL package and a recent change
I4e5440b8450558add372214fd1a0373ab4ad2434.  To unblock the Puppet
OpenStack modules, this change adds a pending repo to the tests.  This
change should be reverted after a newer version of PyMySQL is included
in the RDO packages.

Additionally, we were improperly enabling ipv6 on ubuntu for scenario004
despite the fact the README said it should not be enabled. This change
also disables ipv6 on ubuntu so scenario004 will pass.

Change-Id: I752592a37786b17c5aeb409a048b98f97ca0fc31
Related-Bug: #1641312