Disable puppetlabs-mongodb dbpath_fix
The dbpath_fix in the mongodb module will correct permissions on the
data files. Unfortunately the data files are being created with the
nogroup group so it leads to some idempotency issues when the files get
created.  This is currently blocking the promotion of the module. Since
we don't require this functionality for p-o-i, let's turn it off.

Change-Id: Ia864387f2f4b4bcc02d71c046394d03dae20a411