Reduce the amount of Murano tests from 108 to 61
By including the whole application_catalog regex, we were trying to
run ~108 tempest tests on top of Murano.
Since Murano tempest tests already have the 'smoke' tag, let's just
rely on that to provide the testing that we need.

Murano still has 61 tests tagged as smoke but it's already a start and
we can lower that further by optimizing which tests are run directly
in the Murano tempest plugin afterwards.

Change-Id: I4d975b2c73e1f7fe8013d1ca7638a01b68a24971