Configure api_extensions to test for neutron
If we don't, the default 'all' is used, which means tempest executes all
tests irrespective of whether the setup supports all those extensions,
or not. For example, trunk service plugin is not enabled by the repo;
neither port_security ml2 extension driver is. If a test that requires
any of those features land in tempest repo, the gate will immediately
break. Actually, tempest already has some tests for port_security, but
they are not executed at the moment by the virtue of a special
port_security config option that is False by default. That may change in
the future [1].

It's good practice to explicitly configure the list of extensions that
are known to work for the setup in question.

[1] I471575bdc106e1a6d1ecdb313dbe77b4102d8f15

Change-Id: I323ff03c5798a41de4870ed99e71ed5a9456a6af
Depends-On: I31105600dfd5632ec4341b35ca3e8a13af3bb945
(cherry picked from commit a38b8c9dc336d5c67f308c38dc333313a4380289)