Add workaround for cellv2 race condition
We've seen on ubuntu that the nova computes are taking a bit longer to
register so we end up running discover_hosts prior to the registration.
This results in a failure when we attempt to use nova. So for p-o-i we
need to add in a sleep to give the computes enough time to register
before the discover_hosts is run. In most down stream consumers of these
modules, this should be handled as part of the overall orchestration
(tripleo/fuel/etc) rather than puppet itself because it's something that
will require cross node orchestration.

Change-Id: I9ad0393a34fc1416a3f4959eb83deb7a9b5aa9af