Actually run service validation for designate
The service validation for designate-central was set to refreshonly but
did not have anything actually notifying it to run. This change updates
the validation to listen for the designate::service::end anchor so that
we will wait for the service to be up before trying to manage the pools.

Change-Id: Ife6b29388e97112666ad5602ac62ac7d8534c1d5
Closes-Bug: #1671206
(cherry picked from commit e5d468a42cda7b477418d78a8b67c5471cc223a1)