scenario001: rework telemetry services
Ubuntu and Redhat/Centos doesn't have all telemetry packages.

Instead of checking everywhere if we are on Ubuntu or Redhat/Centos,
this change uses a boolean 'enable_telemetry_legacy'.

If true, ::openstack_integration::ceilometer will configure ceilometer
with legacy services/configuration (without panko/gnocchi). If false,
Gnocchi and Panko backend is used.

scenario001 set it to true for Ubuntu and false for Redhat/Centos.

This centralize the kind of ceilometer deployement in one place.

This change also drops the now useless collector configuration and
configure the notification agent correctly instead.

Change-Id: If229829975d4f25ca0a2a3c6c93e674204d71dc2