Increase swap size
While testing incoming release of eventlet 0.20.1 we've
discovered that it's increasing memory usage for processes
using eventlet. It has been reported upstream [1] but
aparently this is expected.

p-o-i scenario003 is pretty heavy in terms of memory. Before new
eventlet version it's already using ~700-800MB swap in my tests and
with new eventlet (not pushed to RDO Testing repo yet) it has failed
with out-of-memory problems from time to time.

With this patch, i'm increasing swap size from 2GB to 4GB to avoid
oom problems. Note that this may derive in performance issues if
we start having too much swap usage, but i haven't hit it in my
local tests.

[1] https://github.com/eventlet/eventlet/issues/412

Change-Id: If333ae62dbae3f597d2b14ae1dd4c174ecdf51fc