Disable vitrage in scenario001
New service vitrage-collector has been added to vitrage in
https://review.openstack.org/#/c/474560.

To pass tempest tests, vitrage-collector service must be running
so it needs to be added in puppet-vitrage, however this problem
is blocking puppet promotion creating a chicken and egg problem.
This patch disables vitrage deployment temporarily so that we can
fix the issue in puppet-vitrage and get a package with collector
service in promoted repo.

Change-Id: I709347c0570c134cc70e26b75058c7f5c58b3124