Reduce gnocchi metricd cleanup_delay
Because we configure Keystone to expire tokens after 600s, we don't
want to rely on default value in Gnocchi which is 300s to cleanup old data.
Indeed, Gnocchi might use an old token that expired to clean up and then it would
fail. It happens when running Tempest tests in the gate with low resources.
Production value (300) shouldn't be changed by default.

Depends-On: I8d2fccb0cef49246b3c15b0942ba6067519cee1f
Change-Id: If4d4c57698048ea89b9b271e66d6a7a031d21b33