Disable watcher create_schema
According to https://review.openstack.org/#/c/419031/ the only
command to be executed to populate watcher DB is:

watcher-db-manage --config-file /etc/watcher/watcher.conf upgrade head

create_schema command is not longer needed, and in fact is not
creating table apscheduler_jobs properly.

Note that p-o-i jobs keep working because the version of APScheduler
currently in RDO creates table apscheduler_jobs automatically
in SQLAlchemyJobStore initialization. This is not longer true
in recent versions, so it will break as soon as APScheduler is
updated, so let's fix it.

[1] https://github.com/agronholm/apscheduler/commit/bf99bff8ea683b98756313ff34a22320d141186c#diff-c47a8d544459e39b2dc3dfeb50a4ec6bR55

Change-Id: If0338362f9b7b9c90390a836a40b0036adc958a3