Temporarily pin cliff to 2.8.0 in tempest virtualenv
When installing tempest from source we are getting cliff-2.9.0
in the virtualenv. This version of cliff requires that all tempest
plugins have a top_level.txt in the egg-info [1] which is not
present in some RDO packages.

We are fixing this in RDO but it will take a while to fix the
macro, rebuild the packages and get a new puppet promotion
(which currently is blocked by other cinder issue), so I'm
proposing to temporarily pin back to cliff 2.8.0 until RDO
repos are ready for it.

[1] https://review.openstack.org/#/c/501863/

Change-Id: I9c91a316b2f8acaa1f4b0882f0799f90e11464d9