Switch tempest to use RPM by default and disable trove
1) Switch tempest to be installed from RPM instead of source

Currently, tempest is deployed from source by default. This is
problematic as some changes in master trunk can be incompatible
with packages in promoted repo.

This patch switchs p-o-i to deploy tempest from the rpm package
used in the promoted repo, what should fix this kind of scenarios.

2) Disable trove until integration with keystone v3 is fixed upstream

Currently, trove doesn't work fine with keystone API v3. It's
being worked in https://review.openstack.org/#/c/507087/ in
upstream project.

3) Disable tempest tests for object_storage when running with ubuntu
as they are failing with the current version of tempest in the uca repo.

Until it's fixed, let's disable it from p-o-i to unblock promotion.

Change-Id: I385e2f34dd925bdc2fdbc1553d49dce104f566f2