Fix puppet-agent install
The previous declaration where we installed the puppet-agent was after
we revoked sudo access. We need this to be part of the node setup so we
need to move it out of the build run and to the node setup playbooks.

Depends-On: I2fb0301631eb69bd1e37467229bcf3dc4ab532bb
Change-Id: Id68ee1b443a4172d0c1d6d58a04908c52a566623
(cherry picked from commit c586caffa7ef00ae38918a30104f8eb94bfb1392)