Fix puppet-agent install
The previous declaration where we installed the puppet-agent was after
we revoked sudo access. We need this to be part of the node setup so we
need to move it out of the build run and to the node setup playbooks.

Change-Id: Id68ee1b443a4172d0c1d6d58a04908c52a566623
(cherry picked from commit 536643ea9ca23220274d317c63baaf88021e0351)