Empty keystone.conf apache2 vhost
Normally the keystone.conf is removed to make sure the
default configuration provided in the keystone package
for Ubuntu is skipped. This will cause issues when upgrading
the keystone package since the keystone.postinst script when
upgrading will look for that file. [1]

Instead we should make sure this file is empty which will make
sure upgrading the package is possible without manual intervention.

This does not affect gate per se but will act as a guideline on how
it should be used.

[1] https://bugs.launchpad.net/ubuntu/+source/keystone/+bug/1737697

Closes-Bug: #1737697
Change-Id: Ia69a11614a1fdd32a4903b47427fe3e6a48f9205