Remove PUPPET_MAJ_VERSION check for unit modules
The install_modules_unit.sh which is called by the
spec helper to install the modules for the unit
tests checks the PUPPET_MAJ_VERSION variable when
selecting the puppet base path.

The zuul jobs is not configured to give this env
variable so unit tests failed when using the puppet
base path for v3 which was /etc/puppet instead of
/etc/puppetlabs/code that is valid for v4 and v5.

We don't test v3 specifically so the check
should not be needed.

Change-Id: I6a9c53d58f05dfbb8646ad36d3bf5f6bad5588e1