Fix occasional Tempest failures with invalid tokens
The current fernet rotation happens every 5 minutes with a token
expiry of 40 minutes.   This means that after 15 minutes, all
tokens are expired.  This causes Nova to fail for some operations
because it's tokens are now expired.

This adjusts the maximum active keys to 5 and rotation every 10
minutes which will mean that we'll rotate at least once during
the tempest runs which lasts around 15 minutes while keeping all
tokens valid for the entire 40 minutes.

Change-Id: I0e3056aa00a260a93872508dd0158c15060cda19