Remove --regex from tempest command when running with RDO
https://review.openstack.org/#/c/504345/ has changed the behavior of tempest
when running with --regex and --whitelist-file [1]

In order to promote we need to remove this regex option and implement [2] when ready
after promotion.

This patch is removing totally the --regex when not running in UCA. Note these tests
were disabled in [3] and hopefully they are more stable now and we can run
without skipping them until we can implement --blacklist-file.

[1] https://bugs.launchpad.net/tempest/+bug/1751115
[2] https://review.openstack.org/#/c/547278
[3] https://review.openstack.org/#/c/461969/

Change-Id: I0dd846a383ed2e1cb4c699a42e618156a9e628b2