Fix ipv6 certificate to make it compliant with IDNA
Currently we are setting "DNS.0 = ::1", but ::1 is not a valid
A-Label for IDNA so the certificate is not correct.

Additionally, we are setting wrong value for DNS.0 = 127.0.0.1
in the ipv4 certificate.

Finally, removing issuerAltName from both ipv4 and ipv6 certificates
as they are not needed for the jobs.

New versions of python-cryptography are more strict to check
certificates content and does not allow to have not compliant
DNS names so we need to fix the certificate to bump python-cryptography.

Note that horizont tempest plugin does not support ipaddress SANs based
certificate validation so I'm disablint certificate validation for
dashboard in this patch.

Depends-On: Iea7a4b85ac64572fac0f0ad871649a79fbc1c0f5

Change-Id: Ib519d222e07e26d3683b24359e2f67728cdd8029
(cherry picked from commit e40e6d934bb01e0ff1881b3fee92d97d17437179)