Debian: using packages for installing Tempest
Puppet-openstack attempst to install puppet using:

"python setup.py isntall"

Though in Debian, setuptools isn't even installed for Python 2, so it
fails. As the comment in run_tests.sh mentioned the use of setuptools
and pip is temporary, this patch switches to the packaged version of
tempest, testr and os-testr, using apt-get install instead of pip.

Change-Id: I6a2a04dd0e0f2d7d89da538a53af420d52dcfd20