Use latest puppet5 repos
This change makes sure the latest puppet5 version
is used when running integration jobs for puppet5.

The puppetlabs yum mirror is ready to be used and
has the proper packages and is mirrored, example:

* http://mirror.mtl01.inap.openstack.org/yum-puppetlabs/puppet5/

For the puppetlabs apt mirror the puppet5 pool is
needed before this will work. A patch has been fixed
for that here [1]. Therefore this change cannot be
merged until [1] is merged.

[1] https://review.openstack.org/#/c/565867/

Change-Id: I1a2716fdd4b960b12ff24303038bec5ea8fb80a3