Add Ubuntu 18.04 Bionic to CI
We now have support to Mimic and can select that version
and this patch fixes Puppet 5 for Bionic we can add it to CI.

When this is fixed the only thing left is to add beaker acceptance
testing for Bionic.

Bionic only has Puppet 5 and can currently only use Ceph Mimic.

Change-Id: I6f94e214a70f92278f1f3806471ef2f622a938c3