Debian: fix redis service setup
The redis-server appears as "masked" in the OpenStack CI, even though it
did not in my test VM. As a consequence, "systemctl enable redis-server"
fails, and consequently, the puppet run fails in OpenStack CI. The thing
is, there's no need to enable this service, as in all distros, it's
already enabled by default. Even in the OpenStack CI, the service appears
as enabled, but masked (which by the way, should never happen, as this
isn't a working systemd state).

This patch simply removes the enabling of the systemd service, so that
it also works in Debian in the CI.

Change-Id: I1b37524ff1071d87e8a7df2fa3bca66aed162d1b