Debian: no fwaas l3 extensions
Looks like fwaas extensions are making a bunch of functional tests fail
in Debian, probably because of issues in neutron-rpc-server. We disable
it here, at least temporarily, and for Debian only, so that we have a
working tempest run for Debian. This may be reverted later on once a
better solution is found.

Change-Id: I723bb97310dfb2a737ddcdb79caa2a02ef4e211b