Run metadata-json-lint test in lint job
If spec helper is available it will run the rake
task that is included in [1] otherwise it will
install metadata-json-lint and run it manually
like the other lint things does.

Only runs if the metadata.json file is available
so should not require any changes in modules if
something is wrong there.

[1] https://review.openstack.org/591605

Depends-On: https://review.openstack.org/591605
Change-Id: I60571e919e95327bf56a61bc33298a3810b942f7