Fix Ubuntu idempotency and pkg upgrades
The Ubuntu packages for panko-api now defaults
to installing with Apache and WSGI so it drops
the vhost files in the directory which the apache
module then removes.

We force them as empty and by that we keep idempotency
but we also dont break package upgrades for panko-api
because the dh-apache helper for deb packging fails if
these files are removed.

This is the same as we've done for keystone, sahara etc.

Change-Id: I4466d14b825a2392594a1b86b08aeed5793c77b8