Rework the package upgrade after repos
This moves back the generic upgrade process
for packages after the repos are added back
to the run_tests.sh script.

Instead if solves the known python-chardet
dependency bug for requests package for RPM
packaging in the repos.pp manifest.

This is because we need to fix that bug when
running the beaker tests.

We need to keep an eye out if we see any
similar bugs in the future now that updates
are now executed for all packages in beaker
tests.

Change-Id: If12897355d5a4ba7c05e6ee79d63417fcae03977