Add puppet core providers for puppet6
In puppet 6 they have moved out a lot of builtin
resources to their owns modules since they didn't
belong in the Puppet core however in deb/rpm packaging
they are still bundled in but not when installing from
gem.

So we must add them in here so that they are available
in our unit tests.

Change-Id: I7060191395eff9c1e3146cb10594a38bef113117