Set selective rspec-puppet versions
We set the 2.7.1 rspec-puppet version on the
latest non-voting job so that we can start
working on supporting this latest version
were strict variables is enabled.

Since our spec testing does not include classes
in pre_condition properly it fails in later
rspec-puppet versions.

This still leaves our voting jobs with the
2.3.0 versions so that they dont fail.

Change-Id: I7f589d931a948dd9de2100d55f98b95675956768