Limit integration jobs to Puppet5
Because we have so many jobs, it's hard to actually land code. Let's
reduce the gating jobs to just Puppet5 as Puppet 4 is nearing EOL.

This change also moves puppet 6 to experimental if we wish to test it
but does not force puppet 6 on all the jobs. It would be beneficial to
switch the puppet 5 jobs out with puppet 6 jobs next cycle as projects
are migrated to the next version.  TripleO is moving to puppet 5 during
the Stein cycle and tenatively looking to puppet 6 in the Train cycle.

Change-Id: I8ae350a149640badc43abc458c65eb0f93e5fb53