Blacklist tempest test TestTelemetryIntegration
This blacklists the TestTelemetryIntegration
tempest job that is usually failing with the
message [1].

We've had this for a long time and we pretty
much need to recheck every patch, lets blacklist
if for now until somebody can dig up the issue
and solve it.

It's bad since there isn't much telemetry
coverage in Tempest, but there is a autoscaling
job that hopelly tests some aspects for now.

[1] http://logs.openstack.org/19/618919/4/check/puppet-openstack-integration-5-scenario001-tempest-centos-7-luminous/e07a420/job-output.txt.gz#_2019-01-22_16_27_13_049172

Change-Id: Ic6824dafc2fb37168b16f4923cb967e0c8d4947e