Unpin puppet gem version selection
The tickets PUP-9270 [1] and PUP-9271 [2]
is fixed and we should unpin so we use the
latest puppet 5.5 version.

Also changes so that we check that the
templated var puppet_gem_version is actually
containing a value and not just empty if so
we can skip setting the environment variable
and just have the gem unpinned in the spec
helper.

[1] https://tickets.puppetlabs.com/browse/PUP-9270
[2] https://tickets.puppetlabs.com/browse/MODULES-8193

Change-Id: I9b0dbc704638097fc0f6e3702fcea04fb8541b02