Change chardet installation to a exec
The previous code tried to install the python2-chardet
with the source name of python-chardet which for some
reason just straight up skipped doing it.

If you changed the resource name of the package resource
to python2-chardet it failed as well since the python2-chardet
package obsoletes python-chardet, puppet doesn't handle packages
being obsoleted.

Change-Id: Ia5a89d7d716fb08ac75002693dc360bd373903db