Lower neutron rpc max response timeout
This configuration option is used to determine the max exponential
retry which can lead to beaker failures because the ovs agent starts at
the same time as neutron server. We need to force a retry sooner than
waiting 600 seconds for it to try again.

See https://review.openstack.org/#/c/623401/

Depends-On: https://review.openstack.org/#/c/648735/
Change-Id: I7f66c811aa507e76e1b4e087a4a7df5999f4ee50