Replace Ceph Mimic by Nautilus in CentOS
Ceph Nautilus has now a stable release 14.2.0 and it's being packaged
in CentOS Storage SIG. Mimic release was not LTS so let's replace it
by Nautilus where possible.

Since nautilus, the supported store backend is LVM which can use a block
device directly (not loopback) or a LVM logical volume for OSDs.

This patch is creating a LVM structure to create the OSD used to
validate Ceph and switching scenario004 to use nautilus instead of
mimic, so we will keep testing both Luminous and Nautilus, LTS releases.

Depends-On: https://review.opendev.org/#/c/650263/
Change-Id: I81441237a8b7cf651fa6926b2eb3c4ed0d42bb58