Disable strict variables for unit tests
So rspec-puppet 2.4+ has strict variables enabled
which causes test failures. The problem is that the puppet modules
have never worked with strict_variables = true due to backwards
compatibility & import ordering.  Since this has not been true,
lets disable it in the unit testing for now. Perhaps some day
we will be able to turn this back to true.  This should allow us to lift
the 2.3.x cap that we added as part of
https://review.opendev.org/#/c/618854/.

Change-Id: Idbf837e87072c829c94e0356b8536088cf98c3bb