Remove nova::network::neutron::neutron_url
This parameter is deprecated and will be removed in [1].

Removing this will make nova use the default behavior of
looking up the endpoint for talking to neutron. Deployments
that want to override that behavior can set the
neutron_endpoint_override parameter or the use the
neutron_valid_interfaces parameter to let nova know what
type of endpoint type to use.

[1] https://review.opendev.org/#/c/658400/

Change-Id: Id892bf2a2998cb11a5337f36ff55a05dc1f34601