Reparent puppet-openstack-base to legacy-base
Base has been updated to not include thigns like zuul-cloner (which we
depend on). So we need to either remove the usage of this  by using the
repos in /home/zuul/src or set the base to legacy-base so zuul-cloner is
still included. For now let's just reparent the jobs and work on
removing zuul-cloner later.

http://lists.openstack.org/pipermail/openstack-discuss/2019-June/006888.html

Change-Id: Iba69f72b05e70806fc425b72a919d52fd368914d
(cherry picked from commit 2bb40bb6fea5978c24a6e98149176567c4ce9dd1)