Fix checkout for stable/mimic
Since puppet-ceph stable/mimic and puppet-pacemaker stable/0.7.x don't
follow the standard openstack conventions. So when the jobs try and
checkout the code, it is currently getting the master version which
doesn't work as expected. In order to fix this we need to specific the
rocky/mimic branches explicitly in the zuul job config. We did something
similar for stable/jewel in Ie11af4944178fa057a189f4e0c7e4809dc694b86

Change-Id: I3f3bf2b13186d8f3708de0cc7221605d8fc56450