Set Ubuntu beaker jobs to non-voting
All Ubuntu jobs work fine except for the
Beaker tests because they are trying to
use Ceph Nautilus that is not mirrored yet.

This will be fixed in [1]. It's caused by
the bug that is now in [2] because we are
using checkout instead of fixing the branch
and since we are using different versions
of Ceph on jobs we cannot use override-checkout
in Zuul to solve it for now.

[1] https://review.opendev.org/#/c/666263/
[2] https://github.com/openstack/puppet-openstack-integration/blob/master/functions#L42

Change-Id: Iffd1d9abbed52d89bcb5209aac6a1227695f730f