Reduce fernet_rotate frequency
Currently it's every 10 minutes. When using expiration time of 40
minutes (scenario001) or longer, this is too frequent and in some cases,
jobs are failing with InvalidToken in scenario001. To avoid this we can
run fernet_rotate less often or increase max_active_keys [1]. In this
case i'm making it to rotate every 30 minutes instead of 10.

[1] https://docs.openstack.org/keystone/latest/admin/fernet-token-faq.html#i-rotated-keys-and-now-tokens-are-invalidating-early-what-did-i-do

Change-Id: Ia19612b668313752b62545ce5cae49396671b939