Run unittests and lint tests on CentOS8
This patch introduces a job to run unit tests and lint tests on
CentOS8, as we are migrating from CentOS7 to CentOS8.

Note that we have a job with puppet 5.5 initially, because we still
use puppet 5.5 in rdo train, but puppet 5 will be deprecated in Ussri.

Note that we use puppet 6.4 in train so that we use the same version
as ubuntu job.

Conflicts:
	zuul.d/layout.yaml
	zuul.d/unit.yaml
Change-Id: Id643efdbab93f39ce3b168a2d1c451035547ecab
(cherry picked from commit 867f939bd6c71535b49fd6738c225f3815cf5933)