Fix version locks
In order to lock to a specific minor release we need to specify all 3
numbers or it will pick up the next minor release. For 6.11 we were
getting 6.14.0 when the job run which is the same as the latest.

Conflicts:
	zuul.d/libraries.yaml
	zuul.d/unit.yaml

Change-Id: Idab2b1bd50ae316d9d26f5616174a3421f22b571
(cherry picked from commit 92954b3da121573e403077f4067eea0b1087d5e8)