Only run Puppet 6 for integrations
We are finally in Victoria and Puppet 5 is no
longer supported. This removes all Puppet 5
integration testing and moves the Puppet 6
testing from experimental back to check/gate.

This keeps Puppet 5 syntax and unit as a courtesy
until we start introducing Puppet 6 only features
but we don't need to keep backward compatibility
with Puppet 5.

Change-Id: Idc7fcb9538eaa2769a77ec256048d4cef0d1b9ec