Add Puppet Litmus and remove Puppet 5 testing
This patch adds the needed bits to replace Beaker
with Puppet Litmus.

Since Puppet Litmus only supports Puppet 6 we
need to remove Puppet 5 testing. The probability
that we break Puppet 5 support (we have promised
to keep it in Victoria) this late in the cycle is
very low hence we should try to get this through.

When this is merged we can move all modules over
to use Litmus and then cleanup all Beaker things.

Change-Id: I73f18345ed3e6481f86ad7af7d68f66b19d2250e